chore: move certSPKI to laucnher/utils.go
